+++ lib/model/Allergy.dart
@@ -1,0 +1,9 @@
+import 'AllergyName.dart';
+
+class Allergy {
+  num code = 0;
+  String abbreviation = "";
+  AllergyName name = AllergyName("");
+
+  Allergy(this.code, this.abbreviation, this.name);
+}
