+++ lib/model/AllergyName.dart
@@ -1,0 +1,4 @@
+class AllergyName {
+  String ja = "";
+  AllergyName(this.ja);
+}
